private: workarounds for on-device native compiles - force on-device native build environments to use included headers to avoid redefinition errors - fix pointer types for fs_config when using included headers

+++ private/android_filesystem_config.h
@@ -315,7 +315,7 @@ static const struct fs_path_config android_files[] = {
 };
 
 static inline void fs_config(const char *path, int dir,
-                             unsigned *uid, unsigned *gid, unsigned *mode, uint64_t *capabilities)
+                             long unsigned *uid, long unsigned *gid, unsigned *mode, uint64_t *capabilities)
 {
     const struct fs_path_config *pc;
     int plen;
